Move business logic tests to RoutingConfigurerIT
