+++ routing-configurer/src/test/java/no/vegvesen/ixn/federation/qpid/QpidClientIT.java
@@ -1,13 +1,6 @@
 package no.vegvesen.ixn.federation.qpid;
 
-import no.vegvesen.ixn.Sink;
-import no.vegvesen.ixn.Source;
-import no.vegvesen.ixn.federation.api.v1_0.SubscriptionStatus;
 import no.vegvesen.ixn.docker.DockerBaseIT;
-import no.vegvesen.ixn.federation.model.*;
-import no.vegvesen.ixn.ssl.KeystoreDetails;
-import no.vegvesen.ixn.ssl.KeystoreType;
-import no.vegvesen.ixn.ssl.SSLContextFactory;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -23,18 +16,13 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.web.client.RestTemplate;
 import org.testcontainers.containers.GenericContainer;
 
-import javax.jms.JMSException;
-import javax.naming.NamingException;
-import javax.net.ssl.SSLContext;
 import java.net.URL;
-import java.util.*;
+import java.util.LinkedList;
+import java.util.List;
 
-import static java.util.Collections.emptySet;
 import static no.vegvesen.ixn.federation.qpid.QpidClient.FEDERATED_GROUP_NAME;
 import static no.vegvesen.ixn.federation.qpid.QpidClient.SERVICE_PROVIDERS_GROUP_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.fail;
-
 
 @SuppressWarnings("rawtypes")
 @SpringBootTest(classes = {QpidClient.class, QpidClientConfig.class, TestSSLContextConfig.class})
@@ -42,16 +30,10 @@ import static org.junit.Assert.fail;
 @ContextConfiguration(initializers = {QpidClientIT.Initializer.class})
 public class QpidClientIT extends DockerBaseIT {
 
-
 	@ClassRule
 	public static GenericContainer qpidContainer = getQpidContainer("qpid", "jks", "localhost.crt", "localhost.crt", "localhost.key");
 
 	private static Logger logger = LoggerFactory.getLogger(QpidClientIT.class);
-	private static final String NW_EX = "nwEx";
-	private static final String FED_EX = "fedEx";
-	private final SubscriptionRequest emptySubscriptionRequest = new SubscriptionRequest(SubscriptionRequestStatus.EMPTY, emptySet());
-	private final Capabilities emptyCapabilities = new Capabilities(Capabilities.CapabilitiesStatus.UNKNOWN, emptySet());
-	private static String AMQPS_URL;
 	private static Integer MAPPED_HTTPS_PORT;
 
 	static class Initializer
@@ -62,7 +44,6 @@ public class QpidClientIT extends DockerBaseIT {
 			String httpUrl = "http://localhost:" + qpidContainer.getMappedPort(8080);
 			logger.info("server url: " + httpsUrl);
 			logger.info("server url: " + httpUrl);
-			AMQPS_URL = "amqps://localhost:" + qpidContainer.getMappedPort(AMQPS_PORT);
 			TestPropertyValues.of(
 					"qpid.rest.api.baseUrl=" + httpsUrl,
 					"qpid.rest.api.vhost=localhost"
@@ -105,69 +86,6 @@ public class QpidClientIT extends DockerBaseIT {
 	}
 
 	@Test
-	public void interchangeWithOneBindingIsCreated() {
-		Set<Subscription> subscriptions = new HashSet<>(Collections.singletonList(new Subscription("a = b", SubscriptionStatus.REQUESTED)));
-		SubscriptionRequest subscriptionRequest = new SubscriptionRequest(SubscriptionRequestStatus.REQUESTED, subscriptions);
-		Neighbour flounder = new Neighbour("flounder", emptyCapabilities, subscriptionRequest, emptySubscriptionRequest);
-		setupRouting(flounder, NW_EX);
-		assertThat(client.queueExists(flounder.getName())).isTrue();
-	}
-
-	//todo: users of this is candidate for RoutingConfigurerIT
-	private void setupRouting(Subscriber subscriber, String exchangeName) {
-		client.createQueue(subscriber.getName());
-		client.addReadAccess(subscriber.getName(), subscriber.getName());
-		for (Subscription subscription : subscriber.getSubscriptionRequest().getSubscriptions()) {
-			client.addBinding(subscription.getSelector(), subscriber.getName(), subscription.bindKey(), exchangeName);
-		}
-		Set<String> existing = client.getQueueBindKeys(subscriber.getName());
-		Set<String> unwantedBindKeys = subscriber.getUnwantedBindKeys(existing);
-		for (String unwantedBindKey : unwantedBindKeys) {
-			client.unbindBindKey(subscriber.getName(), unwantedBindKey, exchangeName);
-		}
-	}
-
-	@Test
-	public void interchangeWithTwoBindingsIsCreated() {
-		Subscription s1 = new Subscription("a = b", SubscriptionStatus.REQUESTED);
-		Subscription s2 = new Subscription("b = c", SubscriptionStatus.REQUESTED);
-		Set<Subscription> subscriptions = new HashSet<>(Arrays.asList(s1, s2));
-		SubscriptionRequest subscriptionRequest = new SubscriptionRequest(SubscriptionRequestStatus.REQUESTED, subscriptions);
-		Neighbour halibut = new Neighbour("halibut", emptyCapabilities, subscriptionRequest, emptySubscriptionRequest);
-		setupRouting(halibut, NW_EX);
-		assertThat(client.queueExists(halibut.getName())).isTrue();
-	}
-
-	@Test
-	public void interchangeIsBothCreatedAndUpdated() {
-		Set<Subscription> subscriptions = new HashSet<>(Collections.singletonList(new Subscription("a = b", SubscriptionStatus.REQUESTED)));
-		SubscriptionRequest subscriptionRequest = new SubscriptionRequest(SubscriptionRequestStatus.REQUESTED, subscriptions);
-		Neighbour seabass = new Neighbour("seabass", emptyCapabilities, subscriptionRequest, emptySubscriptionRequest);
-		setupRouting(seabass, NW_EX);
-		assertThat(client.queueExists(seabass.getName())).isTrue();
-		setupRouting(seabass, NW_EX);
-		assertThat(client.queueExists(seabass.getName())).isTrue();
-	}
-
-	@Test
-	public void interchangeCanUnbindSubscription() {
-		Subscription s1 = new Subscription("a = b", SubscriptionStatus.REQUESTED);
-		Subscription s2 = new Subscription("b = c", SubscriptionStatus.REQUESTED);
-		Set<Subscription> subscriptions = new HashSet<>(Arrays.asList(s1, s2));
-		SubscriptionRequest subscriptionRequest = new SubscriptionRequest(SubscriptionRequestStatus.REQUESTED, subscriptions);
-		Neighbour trout = new Neighbour("trout", emptyCapabilities, subscriptionRequest, emptySubscriptionRequest);
-		setupRouting(trout, NW_EX);
-		assertThat(client.getQueueBindKeys(trout.getName())).hasSize(2);
-
-		subscriptions = new HashSet<>(Collections.singletonList(s1));
-		subscriptionRequest = new SubscriptionRequest(SubscriptionRequestStatus.REQUESTED, subscriptions);
-		trout = new Neighbour("trout", emptyCapabilities, subscriptionRequest, emptySubscriptionRequest);
-
-		setupRouting(trout, NW_EX);
-		assertThat(client.getQueueBindKeys(trout.getName())).hasSize(1);
-	}
-
-	@Test
 	public void tearDownQueue() {
 		//Set up a new queue
 		client.createQueue("crab");
@@ -200,23 +118,6 @@ public class QpidClientIT extends DockerBaseIT {
 		assertThat(userNames).doesNotContain(deleteUser);
 	}
 
-	@Test
-	public void newServiceProviderCanReadDedicatedOutQueue() throws NamingException, JMSException {
-		ServiceProvider king_gustaf = new ServiceProvider("king_gustaf");
-		setupRouting(king_gustaf, NW_EX);
-		client.addInterchangeUserToGroups(king_gustaf.getName(), SERVICE_PROVIDERS_GROUP_NAME);
-		SSLContext kingGustafSslContext = setUpTestSslContext("jks/king_gustaf.p12");
-		Sink readKingGustafQueue = new Sink(AMQPS_URL, "king_gustaf", kingGustafSslContext);
-		readKingGustafQueue.start();
-		Source writeOnrampQueue = new Source(AMQPS_URL, "onramp", kingGustafSslContext);
-		writeOnrampQueue.start();
-		try {
-			Sink readDlqueue = new Sink(AMQPS_URL, "onramp", kingGustafSslContext);
-			readDlqueue.start();
-			fail("Should not allow king_gustaf to read from queue not granted access on (onramp)");
-		} catch (Exception ignore) { }
-	}
-
 	private static String getFilePathFromClasspathResource(String classpathResource) {
 		URL resource = Thread.currentThread().getContextClassLoader().getResource(classpathResource);
 		if (resource != null) {
@@ -233,7 +134,6 @@ public class QpidClientIT extends DockerBaseIT {
 		assertThat(newACL).hasSize(initialACL.size() + 1);
 	}
 
-
 	@Test
 	public void addOneConsumeRuleBeforeLastRule() {
 		LinkedList<String> acl = new LinkedList<>();
@@ -247,48 +147,6 @@ public class QpidClientIT extends DockerBaseIT {
 		assertThat(acl.getFirst()).isEqualTo(newAcl.getFirst());
 		assertThat(acl.getLast()).isEqualTo(newAcl.getLast());
 		assertThat(newAcl.get(newAcl.size()-2)).contains("king_harald");
-	}
-
-	@Test
-	public void newNeighbourCanNeitherWriteToFedExNorOnramp() throws JMSException, NamingException {
-		HashSet<Subscription> subscriptions = new HashSet<>();
-		subscriptions.add(new Subscription("originatingCountry = 'SE'", SubscriptionStatus.REQUESTED));
-		Neighbour nordea = new Neighbour("nordea", new Capabilities(Capabilities.CapabilitiesStatus.UNKNOWN, emptySet()), new SubscriptionRequest(SubscriptionRequestStatus.REQUESTED, subscriptions), null);
-		setupRouting(nordea, NW_EX);
-		client.addInterchangeUserToGroups(nordea.getName(), FEDERATED_GROUP_NAME);
-		SSLContext nordeaSslContext = setUpTestSslContext("jks/nordea.p12");
-		try {
-			Source writeFedExExchange = new Source(AMQPS_URL, "fedEx", nordeaSslContext);
-			writeFedExExchange.start();
-			writeFedExExchange.send("Ordinary business at the Nordea office.");
-			fail("Should not allow neighbour nordea to write on (fedEx)");
-		} catch (JMSException ignore) {
-		}
-		try {
-			Source writeOnramp = new Source(AMQPS_URL, "onramp", nordeaSslContext);
-			writeOnramp.start();
-			writeOnramp.send("Make Nordea great again!");
-
-			fail("Should not allow nordea to write on (onramp)");
-		} catch (JMSException ignore) { }
-		theNodeItselfCanReadFromAnyNeighbourQueue("nordea");
-	}
-
-
-	/**
-	 * Called from newNeighbourCanWriteToFedExButNotOnramp to avoid setting up more neighbour nodes in the test
-	 * @param neighbourQueue name of the queue to be read by the node itself
-	 */
-	public void theNodeItselfCanReadFromAnyNeighbourQueue(String neighbourQueue) throws NamingException, JMSException {
-		SSLContext localhostSslContext = setUpTestSslContext("jks/localhost.p12");
-		Sink neighbourSink = new Sink(AMQPS_URL, neighbourQueue, localhostSslContext);
-		neighbourSink.start();
-	}
-
-	public SSLContext setUpTestSslContext(String s) {
-		return SSLContextFactory.sslContextFromKeyAndTrustStores(
-				new KeystoreDetails(getFilePathFromClasspathResource(s), "password", KeystoreType.PKCS12, "password"),
-				new KeystoreDetails(getFilePathFromClasspathResource("jks/truststore.jks"), "password", KeystoreType.JKS));
 	}
 
 	@Test
